feat(sidebar-navigation): refactor timer button
Ports timer button components to Typescript and transform to functional
components. Define interfaces for props, cleanup of unused ones and
change the component ui to match the new sidebar navigation style.
