+++ bigbluebutton-html5/imports/ui/components/timer/timer-list-item/styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import Styled from '/imports/ui/components/sidebar-navigation/styles';
+
+const ListItem = styled(Styled.ListItem)``;
+
+export default {
+  ListItem,
+};
